requirements.txt
transformers
torch
pillow
requests
opencv-python
numpy